size_tでキャストしたのを追加

+++ get_next_line.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mshihiro <mshihiro@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: zhihao <zhihao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/17 23:07:44 by zhihao            #+#    #+#             */
-/*   Updated: 2023/03/07 16:42:49 by mshihiro         ###   ########.fr       */
+/*   Updated: 2023/03/08 14:08:28 by zhihao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ char	*read_and_save_file_data(int fd, char *save)
 	char	*buffer;
 	char	*tmp;
 
-	buffer = malloc(sizeof(char) * (BUFFER_SIZE + 1));
+	buffer = malloc(sizeof(char) * (BUFFER_SIZE + size_t(1)));
 	if (!buffer)
 		return (NULL);
 	read_byte = 1;
